Crate zng_ext_setup

Types and service for implementing recoverable install and uninstall tasks.

The primary use for this crate is creating a Windows setup executable that is compressed and packaged using cargo zng res --tool sfx.

Self Setup

For zng apps the installer/uninstaller logic can be build as an alternate mode of the app main executable, avoiding the cost of packaging a separate executable for install/uninstall and sharing the app look and feel.

Crate

This crate is part of the zng project.

Cargo Features

This crate provides 1 feature flag, not enabled by default.

"save"

Enable save/load install config.

Config is saved as ZStd JSON.

Modules

task

Install and uninstall tasks.

Structs

InstallConfig

Represents a list of tasks for a SETUP.install operation.

PreparedInstallConfig

Represents data generated by SETUP.prepare_install that can be used to run a SETUP.commit_install operation.

SETUP

Setup service.

SetupError

Represents a SETUP operation error.

SetupOpStatus

Represents status of a SETUP install or uninstall operation.

SfxClient

Connection with a self-extracting executable generated by cargo zng res --tool sfx.

SfxDataInfo

Represents a data entry that can be served from a SfxClient.

SfxRead

Read requested data.

SfxReadBlocking

Read requested data.

UninstallConfig

Represents data generated by SETUP.install that can be used to run a SETUP.uninstall operation.

Enums

SetupErrorState

Represents state of a setup operation that ended in a SetupError.

SetupStatus

Represents status of SETUP.

SfxError

Error connecting or requesting data using SfxClient.