zng_ext_setup

Module task

Install and uninstall tasks.

See SetupTask docs for a description of the steps an install or uninstall task runs.

Structs

CancelInstallArgs

Arguments for SetupTask::cancel_install.

CopyCurrentExeConfig

Config for CopyCurrentExe.

CreateShortcutConfig

Config for CreateShortcut.

ExtractTarConfig

Config for ExtractTar

InstallArgs

Arguments for SetupTask::install.

InstallTaskError

Error in a SetupTask::install task run.

PrepareInstallArgs

Arguments for SetupTask::prepare_install

TaskTypeId

Unique name for an install or uninstall task.

UninstallArgs

Arguments for SetupTask::uninstall.

ValidateUninstallArgs

Arguments for SetupTask::validate_uninstall.

Enums

CopyCurrentExe

Setup task that copies std::env::current_exe to a location.

CreateShortcut

Setup task that creates a shortcut to an executable file.

ExtractTar

Setup task that extracts TAR container to a new or existing directory on install and removes these files on uninstall.

SetupTaskError

Represents a SetupTask step error.

Traits

SetupTask

Represents an install and uninstall task implementation.